+++ CHANGELOG.md
@@ -7,6 +7,7 @@ DateRange.parse("20250101..20250112).stretch_to_end_of_month
 # => "20250101..20250131"
 ```
 
+  *Vincent Oord*
 
 * Update to Ruby 3.4.6 compatibility